Add gamepad support
The extension now works for PS3 gamepads. All the buttons are supported
however, L2 and R2 lack their analog-ness because most gamepads don't
support that physically.

Add a reporter block for the direction of the analog sticks. Very useful
when you  want to use the sticks to control the direction of a sprite.
The alternative is complicated calculations involving arctangents on the
X and Y coordinates of the sticks.

The coordinates of the sticks are normalized between -100 and 100.
Deadzones are implemented according to "The Right Way" from
[this article][Doing Thumbstick Dead Zones Right].

Rumble functionality has a command block but is still missing.

[Doing Thumbstick Dead Zones Right]: http://www.third-helix.com/2013/04/12/doing-thumbstick-dead-zones-right.html